Moves all PartiQLTypes to their own files
