+++ partiql-types/src/main/kotlin/org/partiql/value/ArrayType.kt
@@ -1,0 +1,11 @@
+package org.partiql.value
+
+/**
+ * PartiQL's Array type
+ *
+ * Aliases include LIST
+ */
+public object ArrayType : PartiQLCoreTypeBase() {
+    override val name: String = "ARRAY"
+    override fun toString(): String = this.name
+}
