+++ partiql-types/src/main/kotlin/org/partiql/value/BagType.kt
@@ -1,0 +1,9 @@
+package org.partiql.value
+
+/**
+ * PartiQL's BAG type
+ */
+public object BagType : PartiQLCoreTypeBase() {
+    override val name: String = "BAG"
+    override fun toString(): String = this.name
+}
