+++ partiql-types/src/main/kotlin/org/partiql/value/BitType.kt
@@ -1,0 +1,8 @@
+package org.partiql.value
+
+public data class BitType(
+    val length: Int
+) : PartiQLCoreTypeBase() {
+    override val name: String = "BIT"
+    override fun toString(): String = "${this.name}(${this.length})"
+}
