+++ partiql-types/src/main/kotlin/org/partiql/value/BitVaryingType.kt
@@ -1,0 +1,8 @@
+package org.partiql.value
+
+public data class BitVaryingType(
+    val length: Int
+) : PartiQLCoreTypeBase() {
+    override val name: String = "BIT_VARYING"
+    override fun toString(): String = "${this.name}(${this.length})"
+}
