+++ partiql-types/src/main/kotlin/org/partiql/value/BoolType.kt
@@ -1,0 +1,5 @@
+package org.partiql.value
+
+public object BoolType : PartiQLCoreTypeBase() {
+    public override val name: String = "BOOL"
+}
