+++ partiql-types/src/main/kotlin/org/partiql/value/ByteType.kt
@@ -1,0 +1,8 @@
+package org.partiql.value
+
+/**
+ * TODO: Should this be allowed? It's not in SQL:1999
+ */
+public object ByteType : PartiQLCoreTypeBase() {
+    override val name: String = "BYTE"
+}
