+++ partiql-types/src/main/kotlin/org/partiql/value/DynamicType.kt
@@ -1,0 +1,6 @@
+package org.partiql.value
+
+public object DynamicType : PartiQLType.Abstract {
+    override val name: String = "DYNAMIC"
+    override fun toString(): String = this.name
+}
