+++ partiql-types/src/main/kotlin/org/partiql/value/Float32Type.kt
@@ -1,0 +1,10 @@
+package org.partiql.value
+
+/**
+ * Approximate Numeric Type
+ *
+ * Aliases include: REAL
+ */
+public object Float32Type : PartiQLCoreTypeBase() {
+    override val name: String = "FLOAT32"
+}
