+++ partiql-types/src/main/kotlin/org/partiql/value/NullType.kt
@@ -1,0 +1,8 @@
+package org.partiql.value
+
+/**
+ * Ion's NULL.NULL type
+ */
+public object NullType : PartiQLCoreTypeBase() {
+    override val name: String = "NULL"
+}
