+++ partiql-types/src/main/kotlin/org/partiql/value/PartiQLCoreTypeBase.kt
@@ -1,0 +1,6 @@
+package org.partiql.value
+
+@Deprecated("How can I make this internal?")
+public abstract class PartiQLCoreTypeBase : PartiQLType.Runtime.Core {
+    override fun toString(): String = this.name
+}
