+++ partiql-types/src/main/kotlin/org/partiql/value/TupleType.kt
@@ -1,0 +1,11 @@
+package org.partiql.value
+
+/**
+ * PartiQL's Tuple type
+ *
+ * Aliases include STRUCT TODO: Are we sure?
+ */
+public object TupleType : PartiQLCoreTypeBase() {
+    override val name: String = "TUPLE"
+    override fun toString(): String = this.name
+}
